[executorch][schema] Add 'EXTERNAL' to DataLocation in schema
To indicate if a tensor is external to the PTE file or not.

Currently, we can also use the existence of 'fqn' to determine if a tensor is external or not.
I think it's better to have a specific location field as fqn may be required for cases besides external tensor storage.

Differential Revision: [D66523171](https://our.internmc.facebook.com/intern/diff/D66523171/)

ghstack-source-id: 256645340
Pull Request resolved: https://github.com/pytorch/executorch/pull/7191